+++ app/error.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { FullHeightContainer } from '@/components/atoms/FullHeightContainer/FullHeightContainer';
-import { ErrorHero } from '@/components/organisms/Heros/Error/Error';
+import { FullHeightContainer } from '@/components/atoms/FullHeightContainer';
+import { ErrorHero } from '@/components/molecules/Error';
 
 const Page = () => (
   <FullHeightContainer size="md">
