+++ app/projects/page.tsx
@@ -1,13 +1,13 @@
-import { Blog } from '@/components/molecules/Blog';
 import { TitleHero } from '@/components/molecules/Heros/Title';
+import { Projects } from '@/components/molecules/Projects';
 import { Container } from '@mantine/core';
 
 const Page = () => (
   <Container size="sm">
     <section>
-      <TitleHero secondaryTitle="software engineering blog" />
+      <TitleHero secondaryTitle="personal projects" />
     </section>
-    <Blog />
+    <Projects />
   </Container>
 );
