+++ components/atoms/Cards/BlogCard/BlogCard.module.css
@@ -1,0 +1,32 @@
+.card {
+  position: relative;
+  overflow: hidden;
+  transition:
+    transform 150ms ease,
+    box-shadow 100ms ease;
+  padding: var(--mantine-spacing-xl);
+  padding-left: calc(var(--mantine-spacing-xl) * 1.5);
+
+  @mixin hover {
+    box-shadow: var(--mantine-shadow-md);
+    transform: scale(1.02);
+  }
+
+  &::before {
+    content: '';
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    width: rem(8px);
+    background-image: linear-gradient(
+      -60deg,
+      var(--mantine-color-violet-8) 0%,
+      var(--mantine-color-grape-7) 100%
+    );
+  }
+}
+
+.dimmed-text {
+  color: var(--mantine-color-dimmed);
+}
