+++ components/atoms/Cards/ExpertiseCard/index.tsx
@@ -1,0 +1,16 @@
+import { ExpertiseCardParams } from '@/types/ExpertiseCardParams';
+import { Card as MantineCard, Title } from '@mantine/core';
+import Image from 'next/image';
+import classes from './Card.module.css';
+
+const ExpertiseCard = ({ title, content, image }: ExpertiseCardParams) => (
+  <MantineCard className={classes['card']}>
+    <div className={classes['title']}>
+      <Image src={image.path} width={35} height={35} alt={image.alt} />
+      <Title order={4}>{title}</Title>
+    </div>
+    {content}
+  </MantineCard>
+);
+
+export { ExpertiseCard };
