+++ components/atoms/CentredImage/CentredImage.module.css
@@ -1,0 +1,8 @@
+.attribution {
+  font-size: var(--mantine-font-size-xs);
+  text-align: center;
+}
+
+.image {
+  margin: 0 auto;
+}
