+++ components/atoms/NavigationButtons/NavigationButtons.module.css
@@ -1,0 +1,19 @@
+.outline-button {
+  color: light-dark(
+    var(--mantine-color-violet-9),
+    var(--mantine-color-violet-6)
+  );
+  border-color: light-dark(
+    var(--mantine-color-violet-9),
+    var(--mantine-color-violet-6)
+  );
+}
+
+.filled-button {
+  background-color: var(--mantine-color-violet-8);
+  color: var(--mantine-color-white) !important;
+}
+
+.filled-button:hover {
+  background-color: var(--mantine-color-violet-9);
+}
