--- types/CardParams.ts
@@ -1,10 +1,0 @@
-interface CardParams {
-  title: string;
-  content: JSX.Element;
-  image: {
-    path: string;
-    alt: string;
-  };
-}
-
-export type { CardParams };
