+++ types/ExpertiseCardParams.ts
@@ -1,0 +1,10 @@
+interface ExpertiseCardParams {
+  title: string;
+  content: JSX.Element;
+  image: {
+    path: string;
+    alt: string;
+  };
+}
+
+export type { ExpertiseCardParams };
