+++ types/ProjectCardParams.ts
@@ -1,0 +1,10 @@
+interface ProjectCardParams {
+  title: string;
+  excerpt: string;
+  progress: number;
+  color: string;
+  icon: 'up' | 'down';
+  link: string;
+}
+
+export type { ProjectCardParams };
